requirements.txt
python-ldap==2.4.15
wsgiref==0.1.2
